Athorization Added

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,12 +1,15 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace RestaurantAPI.Services
@@ -16,12 +19,17 @@ namespace RestaurantAPI.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
 
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger, 
+            IAuthorizationService authorizationService, IUserContextService userContextService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContextService = userContextService;
         }
 
         public RestaurantDto GetById(int id)
@@ -44,6 +52,7 @@ namespace RestaurantAPI.Services
         public int Create(CreateRestaurantDto dto)
         {
             var restaurnat = _mapper.Map<Restaurant>(dto);
+            restaurnat.CreatedById = _userContextService.GetUserId;
             _dbContext.Restaurants.Add(restaurnat);
             _dbContext.SaveChanges();
 
@@ -57,26 +66,36 @@ namespace RestaurantAPI.Services
             var restaurant = _dbContext.Restaurants.FirstOrDefault(r => r.Id == id);
 
             if (restaurant is null) throw new NotFoundException("Restaurant not found");
-            else
+            var authorizationResault = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizationResault.Succeeded)
             {
+                throw new ForbidException();
+            }
+            
                 _dbContext.Restaurants.Remove(restaurant);
                 _dbContext.SaveChanges();
-            }
+            
         }
 
         public void Edit(int id, EditRestaurantDto dto)
         {
+            
             var restaurant = _dbContext.Restaurants.FirstOrDefault(r => r.Id == id);
             
             if (restaurant is null) throw new NotFoundException("Restaurant not found");
-            else
+
+            var authorizationResault =  _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if(!authorizationResault.Succeeded)
             {
+                throw new ForbidException();
+            }
                 restaurant.Name = dto.Name;
                 restaurant.Description = dto.Description;
                 restaurant.HasDelivery = dto.HasDelivery;
 
                 _dbContext.SaveChanges();
-            }
 
         }
     }
